Edit name and func
